Prevent an admin user from deleting himself

+++ app/Http/Controllers/ProfileController.php
@@ -13,51 +13,60 @@ use Inertia\Response;
 
 class ProfileController extends Controller
 {
-    /**
-     * Display the user's profile form.
-     */
-    public function edit(Request $request): Response
-    {
-        return Inertia::render('Profile/Edit', [
-            'mustVerifyEmail' => $request->user() instanceof MustVerifyEmail,
-            'status' => session('status'),
-        ]);
-    }
+	/**
+	 * Display the user's profile form.
+	 */
+	public function edit(Request $request): Response
+	{
+		return Inertia::render('Profile/Edit', [
+			'mustVerifyEmail' => $request->user() instanceof MustVerifyEmail,
+			'status' => session('status'),
+		]);
+	}
 
-    /**
-     * Update the user's profile information.
-     */
-    public function update(ProfileUpdateRequest $request): RedirectResponse
-    {
-        $request->user()->fill($request->validated());
+	/**
+	 * Update the user's profile information.
+	 */
+	public function update(ProfileUpdateRequest $request): RedirectResponse
+	{
+		$request->user()->fill($request->validated());
 
-        if ($request->user()->isDirty('email')) {
-            $request->user()->email_verified_at = null;
+		if ($request->user()->isDirty('email')) {
+			$request->user()->email_verified_at = null;
+		}
+
+		$request->user()->save();
+
+		return Redirect::route('profile.edit');
+	}
+
+	/**
+	 * Delete the user's account.
+	 */
+	public function destroy(Request $request): RedirectResponse
+	{
+		$user = $request->user();
+
+        if ($user->hasRole('ADMIN')) {
+            abort(403, 'Error. For safety reasons, admins are not allowed to delete their accounts.');
+
+			// return Redirect::route('profile.edit')
+			// 	->with('error', 'Admins are not allowed to delete their accounts.');
         }
 
-        $request->user()->save();
+		$request->validate([
+			'password' => ['required', 'current_password'],
+		]);
 
-        return Redirect::route('profile.edit');
-    }
+		$user = $request->user();
 
-    /**
-     * Delete the user's account.
-     */
-    public function destroy(Request $request): RedirectResponse
-    {
-        $request->validate([
-            'password' => ['required', 'current_password'],
-        ]);
+		Auth::logout();
 
-        $user = $request->user();
+		$user->delete();
 
-        Auth::logout();
+		$request->session()->invalidate();
+		$request->session()->regenerateToken();
 
-        $user->delete();
-
-        $request->session()->invalidate();
-        $request->session()->regenerateToken();
-
-        return Redirect::to('/');
-    }
+		return Redirect::to('/');
+	}
 }
